Creando Controllador, validaciones y repositorio para registrar clientes

+++ DTOs/DetalleOrdenCreateDTO.cs
@@ -2,9 +2,9 @@
 {
     public class DetalleOrdenCreateDTO
     {
-        public int OrdenId { get; set; }
+        public long OrdenId { get; set; }
         public OrdenDTO Orden { get; set; }
-        public int ProductoId { get; set; }
+        public long ProductoId { get; set; }
         public ProductoDTO Producto { get; set; }
         public int Cantidad { get; set; }
         public decimal Impuesto { get; set; }
