+++ DTOs/DetalleOrdenDTO.cs
@@ -2,10 +2,10 @@
 {
     public class DetalleOrdenDTO
     {
-        public int DetalleOrdenId { get; set; }
-        public int OrdenId { get; set; }
+        public long DetalleOrdenId { get; set; }
+        public long OrdenId { get; set; }
         public OrdenDTO Orden { get; set; }
-        public int ProductoId { get; set; }
+        public long ProductoId { get; set; }
         public ProductoDTO Producto { get; set; }
         public int Cantidad { get; set; }
         public decimal Impuesto { get; set; }
