+++ DTOs/ProductoDTO.cs
@@ -2,7 +2,7 @@
 {
     public class ProductoDTO
     {
-        public int ProductoId { get; set; }
+        public long ProductoId { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public decimal Precio { get; set; }
