+++ Models/Cliente.cs
@@ -5,7 +5,7 @@ namespace Prueba_Completa_NET.Models
     public class Cliente
     {
         [Key]
-        public int ClienteId { get; set; }
+        public long ClienteId { get; set; }
         public string Nombre { get; set; }
         public string Identidad { get; set; }
         public DateTime CreatedAt { get; set; }
