+++ Models/Producto.cs
@@ -5,7 +5,7 @@ namespace Prueba_Completa_NET.Models
     public class Producto
     {
         [Key]
-        public int ProductoId { get; set; }
+        public long ProductoId { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public decimal Precio { get; set; }
